requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0

# LangChain & LangGraph
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langgraph==0.0.30

# LLM Integration
langchain-google-genai==0.0.6
google-generativeai==0.3.0

# Vector Database & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
tavily-python==0.3.0
aiohttp==3.9.1

# Data Processing  
pandas==2.0.3
openpyxl==3.1.2

# Database
sqlite3  # Built into Python

# Telegram Bot
python-telegram-bot==20.7

# Web API
Flask==3.0.0
Flask-CORS==4.0.0

# Logging & Utilities
colorlog==6.8.0
